+++ ssb_parser/CHANGES.md
@@ -1,5 +1,6 @@
 # v0.1.2
 * Minor performance improvements
+* Allow empty texture value
 
 # v0.1.1
 * Updated documentation