+++ ssb_parser/CHANGES.md
@@ -1,6 +1,9 @@
-# v0.1.2
+# v0.2.0
 * Minor performance improvements
+* Added clone trait to all structures
 * Allow empty texture value
+* Save texture url in Ssb structure
+* Removed search path from parsing parameters
 
 # v0.1.1
 * Updated documentation